requirements.txt
streamlit
tensorflow
huggingface_hub
transformers
scikit-learn
pillow
numpy
pandas 
matplotlib